Add sheets-based hours display
This does a couple of things:

The GSheets auth was formerly contained within SheetsHourCountUpdater,
but I split it off into a SheetsProvider to allow a single Sheets
instance app-wide. (The Sheets class is thread-safe.) Now we have a
LoggedInDisplayer with a refresh function which is called whenever
the makeup might have changed; i.e. clocking in/out, removing a user,
editing a user, voiding a clock, or resetting hours.